more SOD2 fixes

+++ CUE4Parse/GameTypes/SOD2/Assets/Objects/FItemsBitArray.cs
@@ -7,5 +7,5 @@ namespace CUE4Parse.GameTypes.SOD2.Assets.Objects;
 
 public struct FItemsBitArray(FAssetArchive Ar) : IUStruct
 {
-    public BitArray? Data = new BitArray(Ar.ReadArray<byte>(Ar.Read<int>().Align(8) >> 3));
+    public BitArray? Data = new BitArray(Ar.ReadArray<byte>(((int)Ar.Read<ushort>()).Align(8) >> 3));
 }
